raspberry-agent/requirements.txt
paho-mqtt
tflite-runtime
numpy
Pillow
python-dotenv
requests
